requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
zappa==0.60.0